requirements.txt
python-dotenv
openai>=1.40.0
weaviate-client>=4.6.0
redis>=5.0.0
psutil
numpy
pandas
tqdm
scikit-learn
matplotlib
loguru
